requirements.txt
web3==5.30.0
requests==2.28.1
networkx==2.7.1
github3.py==1.3.0
python-dotenv==0.21.0